[receiver/filelog] Add settings to annotate file owner and group names (#30776)
**Description:** When you want to filter log on a shared host with
multiple user, you want to be able to add owner and group of the file
log. Currenly this lookup it's not possible so it's hard to filter
after.
If include_file_infos is true when reading a file on filelogreceiver, it
will add the file owner as the attribute `log.file.owner` and the file
group as the attribute `log.file.group`.
**Link to tracking Issue:** #30775

**Testing:** Add unity test 

**Documentation:** Add documentation on file_input and file log
receiver.